Una pila de cosas

+++ Hoteles Cubanacan/src/com/resline/cubanacan/ui/adapter/DestinosCardAdapter.java
@@ -32,7 +32,7 @@ public class DestinosCardAdapter extends RecyclerViewCardsAdapter {
     }
 
     @Override
-    protected Bundle getBundle(int id) {
+    protected Bundle getBundle(Long id) {
         return new Bundle();
     }
